Replaced the old sequence editor with the new one
