+++ src/components/sequencing/CommandTooltip.svelte
