+++ src/routes/sequencing/edit/[id]/+page.ts
@@ -15,9 +15,11 @@ export const load: PageLoad = async ({ parent, params }) => {
 
     if (sequenceIdAsNumber !== null) {
       const initialSequence: UserSequence | null = await effects.getUserSequence(sequenceIdAsNumber, user);
+      const adaptation = await effects.getSequenceAdaptation(user);
 
       if (initialSequence !== null) {
         return {
+          adaptation,
           initialSequence,
           user,
         };
